+++ nodes_image.py
@@ -13,19 +13,17 @@ from PIL import Image, ImageOps, ImageSequence
 from spandrel import ModelLoader, ImageModelDescriptor
 
 import folder_paths
-from comfy import model_management, utils, samplers
+from comfy import model_management, utils
 from comfy_api.latest import ComfyExtension, io, ui
 from node_helpers import conditioning_set_values, pillow
-from .nodes_vae import vae_encode
 from .nodes_text_encoder import SingleCondCFGGuider
+from .nodes_vae import vae_encode
 
 CATEGORY = "TenserTensor/Image"
 RESIZE_METHODS = ["nearest-exact", "bilinear", "area", "bicubic", "bislerp"]
 FILENAME_FORMATS = ["name-###", "date-name-###", "name-datetime"]
 FORMAT_EXT = {"PNG": ".png", "JPEG": ".jpg", "WEBP": ".webp", }
 MEGAPIXELS = ["0.25 MP", "0.5 MP", "1 MP", "2 MP", "4 MP", "8 MP"]
-
-
 
 
 def align_to_step(size, scale_factor, dimension_step):
@@ -172,8 +170,8 @@ class TT_ImageLoaderResizerNode(io.ComfyNode):
                 io.Combo.Input("image", options=get_image_files(), upload=io.UploadType.image)
             ],
             outputs=[
-                io.Image.Output(display_name="IMAGE"),
-                io.Mask.Output(display_name="MASK"),
+                io.Image.Output("IMAGE"),
+                io.Mask.Output("MASK"),
             ],
         )
 
@@ -332,12 +330,15 @@ def load_upscale_model(upscale_model) -> ImageModelDescriptor:
 BUFFER_FACTOR = 384.0
 
 
-def upscale(timage, **kwargs):
-    device_name, upscale_tile, upscale_overlap = (
-        kwargs.get("upscaler_device", None),
+def upscale(**kwargs):
+    image, device_name, upscale_tile, upscale_overlap = (
+        kwargs.get("image"),
+        kwargs.get("upscaler_device"),
         kwargs.get("upscale_tile"),
         kwargs.get("upscale_overlap")
     )
+
+    timage = image.clone()
     device = get_torch_device(device_name)
     model = load_upscale_model(kwargs.get("upscale_model"))
     i_image = timage.movedim(-1, 1).to(device)
@@ -410,10 +411,11 @@ class TT_ImagePreviewUpscaleSaveNode(io.ComfyNode):
 
     @classmethod
     def execute(cls, **kwargs) -> io.NodeOutput:
-        timage = kwargs.get("image")
+        image = kwargs.get("image")
+        timage = image.clone()
 
         if kwargs.get("upscale_image"):
-            timage = upscale(timage, **kwargs)
+            timage = upscale(**kwargs)
 
         if kwargs.get("save_image"):
             kwargs["image"] = timage